Implement board evaluation function
This function counts the number of possible moves for black and white.
The score of a board configuration for a player is defined as

    (number of moves for me) - (number of moves for opponent)

Each player wants to maximize this score for themselves, which will in
turn minimize it for the other player. As a consequence, this evaluation
function is a zero-sum.